Began refactor of api, to accept multiple inputs
Symbols are stored in an array and each item in the array is run through the function db, which gets the number of likes from the database and price from the iex api. Values are stored in the obj object and pushed to the response array. Response array is returned to the user